Added Pearson+ support
The Pearson+ login system is separate from the reader+/etext one, so it makes sense in terms of UX to separate the two. We will still keep all the relevant code in the prs.py file. Note that psp has many more formats not yet supported by prs, so this commit is just adding the foundation.

Added a disclaimer for scuolabook (adresses issue #50)

Implemented token refresh (for now only in prs/psp)